+++ intelmq/bots/parsers/generic/parser_csv.py
@@ -86,6 +86,10 @@ class GenericCsvParserBot(ParserBot):
             raise ValueError("Length of parameters 'columns' (%d) and 'columns_required' (%d) "
                              "needs to be equal." % (len(self.columns), len(self.columns_required)))
 
+        if self.type:
+            # TODO remove in the next major release
+            self.default_fields["classification.type"] = self.type
+
         self.compose = self.compose_fields or {}
 
     def parse(self, report):
@@ -165,8 +169,6 @@ class GenericCsvParserBot(ParserBot):
         for key, value in self.compose.items():
             event[key] = value.format(*row)
 
-        if self.type is not None and "classification.type" not in event:
-            event.add('classification.type', self.type)
         event.add("raw", self.recover_line(row))
         yield event